+++ crates/nostr/src/prelude.rs
@@ -69,6 +69,8 @@ pub use crate::nips::nip59::{self, *};
 pub use crate::nips::nip65::{self, *};
 pub use crate::nips::nip90::{self, *};
 pub use crate::nips::nip94::{self, *};
+#[cfg(all(feature = "std", feature = "nip96"))]
+pub use crate::nips::nip96::{self, *};
 pub use crate::nips::nip98::{self, *};
 pub use crate::signer::{self, *};
 pub use crate::types::*;